[Refactor] Huge cleanup
