+++ loopa/src/main/java/org/loopa/monitor/MonitorTwitterCreator.java
@@ -1,0 +1,138 @@
+package org.loopa.monitor;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import org.loopa.element.adaptationlogic.AdaptationLogic;
+import org.loopa.element.adaptationlogic.IAdaptationLogic;
+import org.loopa.element.adaptationlogic.enactor.AdaptationLogicEnactor;
+import org.loopa.element.adaptationlogic.enactor.IAdaptationLogicEnactor;
+import org.loopa.element.functionallogic.FunctionalLogic;
+import org.loopa.element.functionallogic.IFunctionalLogic;
+import org.loopa.element.functionallogic.enactor.IFunctionalLogicEnactor;
+import org.loopa.element.functionallogic.enactor.monitor.IMonitorManager;
+import org.loopa.element.functionallogic.enactor.monitor.MonitorManagerTwitter;
+import org.loopa.element.functionallogic.enactor.monitor.MonitorFunctionalLogicEnactor;
+import org.loopa.element.knowledgemanager.IKnowledgeManager;
+import org.loopa.element.knowledgemanager.KnowledgeManager;
+import org.loopa.element.knowledgemanager.adaptiveknowledgemanager.AdaptiveKnowledgeManager;
+import org.loopa.element.knowledgemanager.adaptiveknowledgemanager.IAdaptiveKnowledgeManager;
+import org.loopa.element.logicselector.ILogicSelector;
+import org.loopa.element.logicselector.LogicSelector;
+import org.loopa.element.logicselector.messagedispatcher.ILogicMessageDispatcher;
+import org.loopa.element.logicselector.messagedispatcher.LogicMessageDispatcher;
+import org.loopa.element.messagecomposer.IMessageComposer;
+import org.loopa.element.messagecomposer.MessageComposer;
+import org.loopa.element.messagecomposer.dataformatter.DataFormatter;
+import org.loopa.element.messagecomposer.dataformatter.IDataFormatter;
+import org.loopa.element.messagecomposer.messagecreator.IMessageCreator;
+import org.loopa.element.messagecomposer.messagecreator.MessageCreator;
+import org.loopa.element.receiver.IReceiver;
+import org.loopa.element.receiver.Receiver;
+import org.loopa.element.receiver.messageprocessor.IMessageProcessor;
+import org.loopa.element.receiver.messageprocessor.MessageProcessor;
+import org.loopa.element.sender.ISender;
+import org.loopa.element.sender.Sender;
+import org.loopa.element.sender.messagesender.IMessageSender;
+import org.loopa.element.sender.messagesender.MessageSender;
+import org.loopa.generic.documents.IPolicy;
+import org.loopa.generic.documents.Policy;
+import org.loopa.generic.documents.managers.IPolicyManager;
+import org.loopa.generic.documents.managers.PolicyManager;
+import org.loopa.generic.element.component.ILoopAElementComponent;
+import org.loopa.monitor.IMonitor;
+import org.loopa.monitor.Monitor;
+import org.loopa.comm.message.IMessage;
+import org.loopa.comm.message.Message;
+import org.loopa.kafka.KafkaService;
+
+public class MonitorTwitterCreator {
+  public static IMonitor create(String monitorID, int monFreq) {
+    return new Monitor(monitorID, createReceiver(), createLogicSelector(), createFunctionalLogic(monFreq), createAdaptationLogic(), createMessageComposer(), createSender(), createKnowledgeManager());
+	}
+
+  public static void startMonitoring(IMonitor monitor){
+    HashMap<String, String> hmBodyMessage = new HashMap();
+    hmBodyMessage.put("type", "getMonData");
+    IMessage m = new Message("Main", monitor.getReceiver().getComponentId(), 1, "request", hmBodyMessage);
+    monitor.getReceiver().doOperation(m);
+  }
+
+  private static IReceiver createReceiver(){
+    HashMap hmpReceiver = new HashMap<String, String>();
+    hmpReceiver.put("1", "logicSelectorTwitter");
+    IPolicy rP = new Policy("receiverTwitter", hmpReceiver);
+    IPolicyManager rPM = new PolicyManager(rP);
+    IMessageProcessor rMP = new MessageProcessor();
+    rP.addListerner(rMP);
+    return new Receiver("receiverTwitter", rPM, rMP);
+  }
+
+  private static ILogicSelector createLogicSelector(){
+    HashMap hmpLogicSelector = new HashMap<String, String>();
+    hmpLogicSelector.put("1", "functionalLogicTwitter");
+    IPolicy lsP = new Policy("logicSelectorTwitter", hmpLogicSelector);
+    IPolicyManager lsPM = new PolicyManager(lsP);
+    ILogicMessageDispatcher lsMD = new LogicMessageDispatcher();
+    lsP.addListerner(lsMD);
+    return new LogicSelector("logicSelectorTwitter", lsPM, lsMD);
+  }
+
+  private static IFunctionalLogic createFunctionalLogic(int monFreq) {
+    HashMap hmpFunctionalLogic = new HashMap<String, String>();
+    hmpFunctionalLogic.put("1", "messageComposerTwitter");
+    hmpFunctionalLogic.put("monFreq", Integer.toString(monFreq));
+    IPolicy flP = new Policy("functionalLogicTwitter", hmpFunctionalLogic);
+    IPolicyManager flPM = new PolicyManager(flP);
+    IMonitorManager mm = new MonitorManagerTwitter();
+    IFunctionalLogicEnactor flE = new MonitorFunctionalLogicEnactor(mm);
+    flP.addListerner(flE);
+    return new FunctionalLogic("functionalLogicTwitter", flPM, flE);
+  }
+
+  private static IMessageComposer createMessageComposer() {
+    HashMap hmpMessageComposer = new HashMap<String, String>();
+    hmpMessageComposer.put("1", "senderTwitter");
+    hmpMessageComposer.put("getMonData", "kafkaServiceTwitter");
+    hmpMessageComposer.put("receivedMonData", "kafkaServiceTwitter");
+    IPolicy mcP = new Policy("messageComposerTwitter", hmpMessageComposer);
+    IPolicyManager mcPM = new PolicyManager(mcP);
+    IDataFormatter mcDF = new DataFormatter();
+    IMessageCreator mcMC = new MessageCreator();
+    mcP.addListerner(mcDF);
+    mcP.addListerner(mcMC);
+    return new MessageComposer("messageComposerTwitter", mcPM, mcDF, mcMC);
+  }
+
+  private static ISender createSender() {
+    HashMap hmpSender = new HashMap<String, String>();
+    hmpSender.put("1", "kafkaServiceTwitter");
+    IPolicy sP = new Policy("senderTwitter", hmpSender);
+    IPolicyManager sPM = new PolicyManager(sP);
+    IMessageSender sMS = new MessageSender() {
+      @Override
+      protected void sendMessage(IMessage message) {
+        KafkaService ks = (KafkaService) this.getComponent().getComponentRecipients().get(message.getMessageTo());
+        ks.processRequest(message);
+      }
+    };
+    sP.addListerner(sMS);
+    return new Sender("senderTwitter", sPM, sMS);
+  }
+
+  private static IAdaptationLogic createAdaptationLogic() { // (empty)
+    IPolicy alP = new Policy("adaptationLogicTwitter", new HashMap<String, String>());
+    IPolicyManager alPM = new PolicyManager(alP);
+    IAdaptationLogicEnactor alE = new AdaptationLogicEnactor();
+    alP.addListerner(alE);
+    return new AdaptationLogic("adaptationLogicTwitter", alPM, alE);
+  }
+
+  private static IKnowledgeManager createKnowledgeManager() { // (empty)
+    IPolicy kP = new Policy("knowledgeManagerTwitter", new HashMap<String, String>());
+    IPolicyManager kPM = new PolicyManager(kP);
+    IAdaptiveKnowledgeManager kAKM = new AdaptiveKnowledgeManager();
+    kP.addListerner(kAKM);
+    return new KnowledgeManager("knowledgeManagerTwitter", kPM, kAKM);
+  }
+}
